+++ src/main/java/click/escuela/student/repository/StudentRepository.java
@@ -1,13 +1,16 @@
 package click.escuela.student.repository;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import click.escuela.student.enumerator.GenderType;
 import click.escuela.student.model.Student;
 
 public interface StudentRepository extends JpaRepository <Student, UUID> {
 
 	public List<Student> findBySchool(String school); 
+	public Optional<Student> findByDocumentAndGender(String document,GenderType gender); 
 }
