+++ src/main/java/click/escuela/student/service/impl/StudentServiceImpl.java
@@ -16,7 +16,6 @@ import click.escuela.student.mapper.Mapper;
 import click.escuela.student.model.Student;
 import click.escuela.student.repository.StudentRepository;
 import click.escuela.student.service.ServiceGeneric;
-import click.escuela.student.util.Validation;
 
 @Service
 public class StudentServiceImpl implements ServiceGeneric<StudentApi, StudentDTO> {
@@ -28,7 +27,7 @@ public class StudentServiceImpl implements ServiceGeneric<StudentApi, StudentDTO
 	@Override
 	public void create(StudentApi studentApi) throws TransactionException {
 		
-		//if(!validation.StudentExists(studentApi)) {
+		if(!StudentExists(studentApi)) {
 			try {
 				
 				Student student = Mapper.mapperToStudent(studentApi);
@@ -36,12 +35,12 @@ public class StudentServiceImpl implements ServiceGeneric<StudentApi, StudentDTO
 			} catch (Exception e) {
 				throw new TransactionException(StudentEnum.CREATE_ERROR.getCode(),
 						StudentEnum.CREATE_ERROR.getDescription());
-	//	}
+	}
 		}
-		/*	else {
+			else {
 				throw new TransactionException(StudentEnum.STUDENT_EXIST.getCode(),
 						StudentEnum.STUDENT_EXIST.getDescription());
-			}*/
+			}
 		}
 
 	
@@ -59,7 +58,7 @@ public class StudentServiceImpl implements ServiceGeneric<StudentApi, StudentDTO
 	
 	public void update(String id, StudentApi studentApi) throws TransactionException {	
 		UUID idReal= UUID.fromString(id);
-		if(findID(idReal).equals(null)) {
+		if(!StudentFindID(idReal)) {
 			throw new TransactionException(StudentEnum.UPDATE_ERROR.getCode(),
 					StudentEnum.UPDATE_ERROR.getDescription());
 		}
@@ -99,14 +98,30 @@ public class StudentServiceImpl implements ServiceGeneric<StudentApi, StudentDTO
 		return Mapper.mapperToStudentsDTO(studentRepository.findAll());
 	}
 
-	public Student findID(UUID id) {
+	public boolean StudentFindID(UUID id) {
+		Boolean exist = false;
 		Optional<Student> optional= studentRepository.findById(id);
 		if(optional.isPresent()) {
-			return optional.get();
+			exist=true;
 		}
 		else {
-			return null;
+			exist=false;
 		}
+		return exist;
+	}
+	
+	public boolean  StudentExists(StudentApi student) {
+		Boolean exist = false;
+		
+			Optional<Student> studentExist=studentRepository.findByDocumentAndGender(student.getDocument(), Mapper.mapperToEnum(student.getGender()));
+			if(studentExist.isPresent()) {
+				exist=true;
+			}
+			else {	
+				exist=false;
+				}
+		
+			return exist;
 	}
 	
 	@Override
